[PATCH] enable toolchain-package to be root only when required
Applied this one manually. Original author: Ron <ron@debian.org>

Description as follows:
The subject mostly says it all, this adds $BUILD_ROOTCMD and
$INSTALL_ROOTCMD to the toolchain-package script and uses them
to get only the (fake or real) root privs it needs, when it
needs them.

I used this to build a full uclibc toolchain yesterday, but the
patch has been rebased to apply to your most recent head from a
few minutes ago.  Nothing you changed seems to conflict with it.
Signed-off-by: Alexander Shishkin <virtuoso@slind.org>